# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PUCPR Projeto Final - Webservices e MBaaS ",
        "description": "API criada como projeto final da matéria de Webservices e MBaaS - por Lucas Cândido dos Santos",
        "contact": {
            "name": "Lucas Cândido dos Santos"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "API server"
        }
    ],
    "paths": {
...
        "/principais-marcas": {
            "get": {
                "tags": [
                    "marcas"
                ],
                "summary": "Retorna uma lista com as principais marcas",
                "description": "Retorna uma lista com as principais marcas",
                "operationId": "618c28c96b3ad55ec387fa6cbc57d00e",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "Missing Data"
                    }
                }
            }
        },
        "/marcas": {
            "get": {
                "tags": [
                    "marcas"
                ],
                "summary": "Retorna uma lista com todas as marcas",
                "description": "Retorna uma lista com todas as marcas",
                "operationId": "ae1229fe3ceeb69c84d84348e9a42ed8",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "Missing Data"
                    }
                }
            },
            "post": {
                "tags": [
                    "marcas"
                ],
                "summary": "Cadastra uma nova marca",
                "description": "Cadastra uma nova marca",
                "operationId": "5bd8e50c970263852b63ef439eabcede",
                "parameters": [
                    {
                        "name": "nome",
                        "in": "query",
                        "description": "Nome da marca",
                        "required": true
                    },
                    {
                        "name": "pai",
                        "in": "query",
                        "description": "ID da marca, em caso de ser uma submarca",
                        "required": false
                    },
                    {
                        "name": "imagem",
                        "in": "query",
                        "description": "URL da imagem",
                        "required": false
                    },
                    {
                        "name": "principal",
                        "in": "query",
                        "description": "Define se uma marca é principal ou não",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Não autorizado. Apenas administradores podem cadastrar produtos"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/marcas/{id}": {
            "get": {
                "tags": [
                    "marcas"
                ],
                "summary": "Retorna uma marca via ID",
                "description": "Retorna uma marca via ID",
                "operationId": "41e4c5bd578c622566574c8a7e6f76ef",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Informe o ID da marca desejada",
                        "required": true,
                        "example": "123"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "Missing Data"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "marcas"
                ],
                "summary": "Atualiza uma marca",
                "description": "Atualiza uma marca",
                "operationId": "1be54f200e94eb246450fed82c091d5b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Informe o ID da marca desejada",
                        "required": true,
                        "example": "123"
                    },
                    {
                        "name": "nome",
                        "in": "query",
                        "description": "Nome da marca",
                        "required": false
                    },
                    {
                        "name": "pai",
                        "in": "query",
                        "description": "ID da marca, em caso de ser uma submarca",
                        "required": false
                    },
                    {
                        "name": "imagem",
                        "in": "query",
                        "description": "URL da imagem",
                        "required": false
                    },
                    {
                        "name": "principal",
                        "in": "query",
                        "description": "Define se uma marca é principal ou não",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Não autorizado. Apenas administradores podem cadastrar produtos"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/marcas-categoria/{id}": {
            "get": {
                "tags": [
                    "marcas"
                ],
                "summary": "Retorna uma lista de marcas pertencentes na categoria informada como parâmetro",
                "description": "Retorna uma lista de marcas pertencentes na categoria informada como parâmetro",
                "operationId": "5ab7b7c6f1fc5aeeff64b733f080cb91",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Informe o ID da categoria desejada",
                        "required": true,
                        "example": "123"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "Missing Data"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}